relax bounds of wai-middleware-validation dependencies
